requirements.txt
Flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
sentence-transformers==2.2.2
weaviate-client==3.23.2
mistralai==0.0.7
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
markdown==3.5
PyPDF2==3.0.1
nltk==3.8.1 